requirements.txt
django>=4.0.0,<4.2.0
djangorestframework
requests
python-dotenv==0.21.0